+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 - Fix missing Resume button (#421)
 - When decoding inbound emails, decode encoded headers (#420)
 - Escape markdown in MS Teams notifications (#426)
+- Set the "title" and "summary" fields in MS Teams notifications (#435)
 
 ## v1.16.0 - 2020-08-04